% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessNegs.R
\name{assessNegs}
\alias{assessNegs}
\title{Assess negative controls}
\usage{
assessNegs(asvtable = NULL, ...)
}
\arguments{
\item{asvtable}{An ASV table}

\item{...}{inherited from identifyNegs}
}
\value{
? not sure yet
}
\description{
Assess negative controls
}
\examples{
asvtable <- makeExampleSeqtab()
metadata <- makeExampleMeta()
assessNegs(asvtable, metadata, "SampleType", "negative control")
}
