% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badData.R
\docType{data}
\name{badasv1}
\alias{badasv1}
\title{Example ASV count table that will give warnings on sanity check}
\format{
\subsection{\code{badasv1}}{

A data frame with 12 rows and 10 columns:
\describe{
\item{ID}{Sample ID}

...
}
}
}
\source{
made up data following typical dada2 ASV count table style but with a non-standard SampleID column
}
\usage{
badasv1
}
\description{
Made up ASV count table
}
\keyword{datasets}
