% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badData.R
\docType{data}
\name{badasv2}
\alias{badasv2}
\title{Example ASV count table that will give warnings on sanity check}
\format{
\subsection{\code{badasv2}}{

A data frame with 12 rows and 8 columns:
\describe{
\item{SampleID}{Sample ID}

...
}
}
}
\source{
made up data following typical dada2 ASV count table style but with two ASVs deleted
}
\usage{
badasv2
}
\description{
Made up ASV count table with two ASVs deleted
}
\keyword{datasets}
