% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badData.R
\docType{data}
\name{badasv3}
\alias{badasv3}
\title{Example ASV count table that will give warnings on sanity check}
\format{
\subsection{\code{badasv3}}{

A data frame with 12 rows and 10 columns:
\describe{
\item{SampleID}{Sample ID}

...
}
}
}
\source{
made up data following typical dada2 ASV count table style but with one of the ASV names messed up
}
\usage{
badasv3
}
\description{
Made up ASV count table with one ASV name deliberately messed up
}
\keyword{datasets}
