% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badData.R
\docType{data}
\name{badmetadata1}
\alias{badmetadata1}
\title{Example metadata with different SampleIDs}
\format{
\subsection{\code{badmetadata1}}{

A data frame with 12 rows and 6 columns:
\describe{
\item{SampleID}{Sample ID}
\item{SampleType}{Sample Type}
\item{HealthStatus, Age, Sex, Location}{demographic data on participants}

...
}
}
}
\source{
made up data following typical dada2 ASV count table style but with one of the ASV names messed up
}
\usage{
badmetadata1
}
\description{
Made up metadata object with the wrong Sample IDs
}
\keyword{datasets}
