% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBetaDiv.R
\name{calcBetaDiv}
\alias{calcBetaDiv}
\title{Calculate beta diversity dissimilarities with or without rarefaction}
\usage{
calcBetaDiv(asvtable, numRare = 400, method = "bray")
}
\arguments{
\item{asvtable}{Input ASV table. Can be filtered with filtering()}

\item{numRare}{Number of rarefactions to perform. Set to 0 if you do not want to do any rarefication.}

\item{method}{Dissimilarity index to use. See ?vegan::vegdist for more information.}
}
\value{
A dissimilarity matrix
}
\description{
Calculate beta diversity dissimilarities with or without rarefaction
}
\examples{
asvtable <- makeExampleSeqtab()
calcBetaDiv(asvtable, 400, "bray")
}
