% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAll.R
\name{checkAll}
\alias{checkAll}
\title{Title}
\usage{
checkAll(x, ...)
}
\arguments{
\item{x}{Objects to check (either a single list) or three objects}

\item{...}{can take a flexible number of arguments}
}
\value{
something
}
\description{
Title
}
\examples{
checkAll(exampleData)
checkAll(exampleMetadata, exampleASVtable, exampleTaxa)
}
