% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkASV.R
\name{checkASV}
\alias{checkASV}
\title{Check your ASV count and taxonomy tables for potential issues.}
\usage{
checkASV(asvtable, taxa, metadata)
}
\arguments{
\item{asvtable}{The ASV table}

\item{taxa}{Its associated taxonomy table}

\item{metadata}{The associated metadata table}
}
\value{
Potential warnings or errors.
}
\description{
Check your ASV count and taxonomy tables for potential issues.
}
\examples{
asvtable <- makeExample("asv")
taxa <- makeExample("taxa")
metadata <- makeExample("meta")
checkASV(asvtable, taxa, metadata)
}
