% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2TaxaNames.R
\name{dada2TaxaNames}
\alias{dada2TaxaNames}
\title{Returns a character vector of expected column names for dada2 taxonomy object}
\usage{
dada2TaxaNames()
}
\value{
Returns a character vector of expected column names for dada2 taxonomy object
}
\description{
Not intended to be used by end user; is called upon by other functions.
}
\examples{
dada2TaxaNames()
}
