% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleASVtable}
\alias{exampleASVtable}
\title{Example ASV count table object}
\format{
\subsection{\code{exampleASVtable}}{

A data frame with 12 rows and 10 columns:
\describe{
\item{SampleID}{Sample ID}

...
}
}
}
\source{
made up data following typical dada2 ASV count table style
}
\usage{
exampleASVtable
}
\description{
Made up ASV count table
}
\keyword{datasets}
