% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleMetadata}
\alias{exampleMetadata}
\title{Example metadata object}
\format{
\subsection{\code{exampleMetadata}}{

A data frame with 12 rows and 6 columns:
\describe{
\item{SampleType}{Sample Type}
\item{HealthStatus, Age, Sex, Location}{demographic data on participants}

...
}
}
}
\source{
made up data based on common study design
}
\usage{
exampleMetadata
}
\description{
Made up demographic, clinical, and lab data on samples
}
\keyword{datasets}
