% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleTaxa}
\alias{exampleTaxa}
\title{Example ASV taxonomy table}
\format{
\subsection{\code{exampleTaxa}}{

A data frame with 9 rows and 8 columns:
\describe{
\item{ASV}{ASV name/sequence}
\item{Kingdom, Phylum, Class, Order, Family, Genus, Species}{taxonomic classification at each level}

...
}
}
}
\source{
made up data following typical dada2 ASV taxonomy table style
}
\usage{
exampleTaxa
}
\description{
Made up ASV taxonomy table
}
\keyword{datasets}
