% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filtering}
\alias{filtering}
\title{Filter the ASV table}
\usage{
filtering(seqtab, minDepth, minASVCount)
}
\arguments{
\item{seqtab}{Input ASV table}

\item{minDepth}{Minimum number of reads a sample must have to be kept}

\item{minASVCount}{Minimum count an ASV must have to be kept}
}
\value{
A filtered ASV table
}
\description{
Filter the ASV table
}
