% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyNegs.R
\name{identifyNegs}
\alias{identifyNegs}
\title{Add a column to metadata object which identifies lab negative controls}
\usage{
identifyNegs(df = NULL, column_to_look_in = NULL, how = NULL)
}
\arguments{
\item{df}{Your metadata object}

\item{column_to_look_in}{What is the name of the column that identifies which controls are negative?}

\item{how}{What pattern should be looked for in the column to find the negative controls in that column?}
}
\value{
An updated metadata object with a column called isNeg identifying lab negative controls
}
\description{
Add a column to metadata object which identifies lab negative controls
}
\examples{
metadata <- makeExampleMeta()
identifyNegs(metadata, "SampleType", "negative control")  # "how" can be a vector e.g., "c("negative", "neg control", "negative control")
}
