% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isASVtable.R
\name{isASVtable}
\alias{isASVtable}
\title{Is an object an ASV table?}
\usage{
isASVtable(x)
}
\arguments{
\item{x}{A data frame or tibble}
}
\value{
TRUE if it thinks object is an ASV table, FALSE otherwise
}
\description{
This function tries to identify if a data frame or tibble is an ASV table or not. It is called by other functions, and isn't intended to be used by end user.
}
\examples{
isASVtable(exampleMetadata) # returns FALSE
isASVtable(exampleASVtable) # returns TRUE
}
