% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isMeta.R
\name{isMeta}
\alias{isMeta}
\title{Is an object a metdata table?}
\usage{
isMeta(x)
}
\arguments{
\item{x}{A data frame or tibble}
}
\value{
TRUE if it thinks object is a metadata table, FALSE otherwise
}
\description{
This function tries to identify if a data frame or tibble is a metadata table or not. It is called by other functions, and isn't intended to be used by end user.
}
\examples{
isMeta(exampleMetadata) # returns TRUE
isMeta(exampleASVtable) # returns FALSE
}
