% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTaxa.R
\name{isTaxa}
\alias{isTaxa}
\title{Is an object a taxonomy table?}
\usage{
isTaxa(x)
}
\arguments{
\item{x}{A data frame or tibble}
}
\value{
TRUE if it thinks object is a taxonomy table, FALSE otherwise
}
\description{
This function tries to identify if a data frame or tibble is a taxonomy table or not. It is called by other functions, and isn't intended to be used by end user.
}
\examples{
isTaxa(exampleMetadata) # returns FALSE
isTaxa(exampleASVtable) # returns FALSE
isTaxa(exampleTaxa) # returns TRUE
}
