% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBadExample.R
\name{makeBadExample}
\alias{makeBadExample}
\title{Title}
\usage{
makeBadExample(...)
}
\arguments{
\item{...}{inherited from something tbd}
}
\value{
An example of a "bad" metadata, ASV table, or taxonomy object
}
\description{
Title
}
\examples{
makeBadExample("meta")
}
