% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBadExampleASV.R
\name{makeBadExampleASV}
\alias{makeBadExampleASV}
\title{Make a "bad" ASV count object that will fail checks}
\usage{
makeBadExampleASV(operation)
}
\arguments{
\item{operation}{"ids", "remove", or "rename"}
}
\value{
A "bad" ASV count object that will fail checks
}
\description{
Make a "bad" ASV count object that will fail checks
}
\examples{
makeBadExampleASV("ids")
makeBadExampleASV("remove")
makeBadExampleASV("replace")
}
