% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBadExampleMeta.R
\name{makeBadExampleMeta}
\alias{makeBadExampleMeta}
\title{Make a "bad" metadata object that will fail checks}
\usage{
makeBadExampleMeta(operation)
}
\arguments{
\item{operation}{"ids", "dups", or "wrongmeta"}
}
\value{
A "bad" metadata object that will throw up warnings on sanity checks
}
\description{
Make a "bad" metadata object that will fail checks
}
\examples{
makeBadExampleMeta("ids")
makeBadExampleMeta("dups")
makeBadExampleMeta("wrongmeta")
}
