% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExample.R
\name{makeExample}
\alias{makeExample}
\title{Create an example object}
\usage{
makeExample(what = NULL)
}
\arguments{
\item{what}{What do you want to create? Valid options are meta (for a metadata table), asv (for an ASV table), or taxa (for a taxonomy table)}
}
\value{
Either a metadata, ASV, or taxonomy table, or a list of all three.
}
\description{
Create an example object
}
\examples{
metadata <- makeExample("meta")
asvtable <- makeExample("asv")
taxa     <- makeExample("taxa")
all      <- makeExample().   # OR:
all      <- makeExample("all")
}
