% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExampleMeta.R
\name{makeExampleMeta}
\alias{makeExampleMeta}
\title{Creats a simple metadata table}
\usage{
makeExampleMeta()
}
\value{
Simple metadata table for testing and demonstration.
}
\description{
Creats a simple metadata table
}
\examples{
metadata <- makeExampleMeta()
}
