% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packItUp.R
\name{packItUp}
\alias{packItUp}
\title{Package up your three objects (metadata, asvtable, taxonomy table) into a single list for less typing}
\usage{
packItUp(metadata = NULL, asvtable = NULL, taxa = NULL)
}
\arguments{
\item{metadata}{Metadata object}

\item{asvtable}{ASV count table object}

\item{taxa}{Taxonomy table object}
}
\value{
A list containing the above three objects
}
\description{
Package up your three objects (metadata, asvtable, taxonomy table) into a single list for less typing
}
\examples{
packItUp(makeExample()$metadata, makeExample()$asvtable, makeExample()$taxa)
}
