% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamePackedList.R
\name{renamePackedList}
\alias{renamePackedList}
\title{Title}
\usage{
renamePackedList(x)
}
\arguments{
\item{x}{tba}
}
\description{
Not intended to be used by end user
}
\examples{
renamePackedList(exampleData)
}
