% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switcher.R
\name{switcher}
\alias{switcher}
\title{Title}
\usage{
switcher(x, ...)
}
\arguments{
\item{x}{list or set of arguments}

\item{...}{flexible}
}
\description{
Not intended to be used by end user
}
\examples{
switcher(exampleData)
}
